# [bugfix] Fix access beyond IP layer bounds if packet stops right after it