--- a/pkg/capture/flow.go
+++ b/pkg/capture/flow.go
@@ -71,7 +71,7 @@ func ParsePacket(ipLayer capture.IPLayer) (epHash capturetypes.EPHash, isIPv4 bo
 	var protocol byte
 	if ipLayerType := ipLayer.Type(); ipLayerType == ipLayerTypeV4 {
 
-		_ = ipLayer[ipv4.HeaderLen] // bounds check hint to compiler
+		_ = ipLayer[ipv4.HeaderLen-1] // bounds check hint to compiler
 
 		isIPv4, protocol = true, ipLayer[9]
 
@@ -126,7 +126,7 @@ func ParsePacket(ipLayer capture.IPLayer) (epHash capturetypes.EPHash, isIPv4 bo
 		}
 	} else if ipLayerType == ipLayerTypeV6 {
 
-		_ = ipLayer[ipv6.HeaderLen] // bounds check hint to compiler
+		_ = ipLayer[ipv6.HeaderLen-1] // bounds check hint to compiler
 
 		protocol = ipLayer[6]