Fixed potential out-of-bounds access and added test for the encountered special (invalid) IP packet scenario.

I'm also trying to capture an actual small `pcap` file on the host in question to get a sample for the E2E tests, but didn't happen again so far... :see_no_evil: 

Performance is actually slightly better (probably because the hint now is more accurate for everything that happens in the function):
```
                                                 │ /tmp/old.txt │            /tmp/new.txt            │
                                                 │    sec/op    │   sec/op     vs base               │
Population/0.0.0.0:68->255.255.255.255:67_17_0-4    9.681n ± 1%   9.514n ± 1%  -1.73% (p=0.000 n=25)
```

Nice catch!


Closes #279